update param tuning scripts
separate the ones that run the time series from the spectral methods
